Linear search

+++ 0x1E-search_algorithms/0-linear.c
@@ -12,13 +12,13 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	unsigned int i = 0;
+	int i = 0;
 
 	if (array == NULL || size == 0)
 	{
 		return (-1);
 	}
-	for (i = 0; i < size; i++)
+	for (i = 0; i < (int)size; i++)
 	{
 		printf("Value checked array[%u] = [%d]\n", i, array[i]);
 		if (array[i] == value)
